Expect environment to provide npm
Version 3.10.7 of npm transitively depends on vulnerable packages. Best
to remove it since it doesn't serve any function. The kr8sswordz app is
expected to build in a specific environment (as defined in the
Dockerfile) so we can reasonably expect that npm will be provided by the
environment.

Related to KC-163